requirements.txt
asgiref==3.8.1
crispy-bootstrap4==2024.10
Django==5.1.1
django-crispy-forms==2.3
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
Markdown==3.7
PyJWT==2.9.0
pytz==2024.1
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
